LayoutTests/fast/shadow-dom/cannot-hit-test-into-shadow-textarea-expected.txt

PASS range.commonAncestorContainer is textareaElement
PASS successfullyParsed is true

TEST COMPLETE
